Breaking each function into it's own file
